Brings in 270 commits from main: Deepgram cloud transcription (ADR
0047/0048), MCP tool connectors incl. OAuth (ADR 0051), fractional
frame rates (default 0.5 fps), the macOS capture-suspension model for
unexpected stream stops, Lucide icons via unplugin-icons, the
onboarding permissions factory, and the release.mnema.day updater feed.

Conflict resolutions of note:
- audio-transcription manifest: kept the platform-gated `models` vec
  (Windows hides Apple providers + full-precision Parakeet) and
  appended main's two Deepgram descriptors.
- desktop_processing_registry: takes main's `deepgram_auth_status`
  param AND returns this branch's `(registry, gpu_state)` tuple; both
  managed states are installed.
- lifecycle.rs: adopted main's `suspend_screen_after_unexpected_stop`
  path in tick_inactivity; kept both cfg-disjoint test modules
  (windows transient-liveness + macos suspension).
- Icon.svelte deleted (main); the gpuAcceleration glyph maps to
  lucide/cpu in section-icons.ts; IconName now comes from groups.ts.
- Onboarding: ported the Windows `permissionPermitsCapture` count into
  main's new onboarding-permissions.svelte.ts; main's new
  stepper/jump-to-blocker code uses the platform-filtered feature list.
- updater endpoint: main's release.mnema.day + this branch's Windows
  passive installMode.

Semantic (compile-level) merge fixes:
- Ported fractional f64 frame rates through the Windows capture stack:
  WGC frame cap (frame_schedule.rs), MF sink-writer frame-rate ratio
  (round(rate*1000)/1000), default-bitrate math, win_smoke example,
  and the desktop segment-start path. 0.5 fps now works on Windows
  instead of truncating to 0.
- Windows metadata refresh publishes through main's snapshot-history
  ring (publish_snapshot) instead of the removed latest_snapshot field.
- capture_retention test uses main's CaptureDb::single constructor.
- Gated 8 of main's new MCP tests to macOS: they dial real `sh -c`
  fixtures and/or hit the platform keychain, which is macOS-only
  (SUPPORTS.md marks MCP "Windows unaddressed"; the
  MNEMA_MCP_SERVER_SECRET_DIR fallback is cfg(test)-internal to
  app-infra and compiled out for these tests).

Verified on Windows: cargo check --workspace --all-targets clean;
mnema --lib 905/905, app-infra 633/633, capture crates green;
svelte-check 0 errors; bun test 743/744 (the one failure spawns node,
which this box lacks — passes in CI).

+++ apps/desktop/src-tauri/src/native_capture/runtime.rs
@@ -122,7 +122,7 @@ pub struct NativeCaptureRuntime {
     #[cfg(any(target_os = "macos", target_os = "windows"))]
     pub current_segment_output_files: Option<CaptureOutputFiles>,
     pub current_segment_index: u64,
-    pub screen_frame_rate: u32,
+    pub screen_frame_rate: f64,
     pub screen_resolution: ScreenResolution,
     pub effective_screen_bitrate_bps: Option<u32>,
     pub microphone_device_id_for_capture: Option<String>,
